Global Dark Mode
Global Dark Mode Button is in the profile page now. Got rid of all the other dark mode toggle buttons on every page. Added dark mode to the game page, game result page, and the profile page too. (Except the friend page, because we don't know if we are going to keep this page yet).